requirements.txt
Flask==3.0.3
gunicorn==22.0.0
google-cloud-storage==1.42.3
Werkzeug==3.0.3
